requirements.txt
aiohttp==3.8.4
aiomysql==0.1.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
bson==0.5.10
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
databases==0.7.0
fast-api-gcs==0.0.11
fastapi==0.95.1
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.19.1
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
multidict==6.0.4
mysql-connector-python==8.0.33
mysqlclient==2.1.1
openai==0.27.6
passlib==1.7.4
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
PyMySQL==1.0.3
python-dateutil==2.8.2
python-multipart==0.0.6
requests==2.30.0
rsa==4.9
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.26.1
tqdm==4.65.0
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.26.16
uuid==1.30
uvicorn==0.22.0
yarl==1.9.2
